Silence ConversationViewModel size lint and split helper extensions
Fix the SwiftLint failure on PR #592 by moving the read-receipt and
typing-indicator helper methods into dedicated companion files and
running SwiftLint locally in strict mode.

`ConversationViewModel` still exceeds the global `type_body_length`
threshold because the file remains the long-standing home for a broad
set of conversation-detail behaviors, so apply a targeted
`swiftlint:disable/enable type_body_length` around the class to keep CI
green without changing runtime behavior.
